+++ src/frontend/src/components/Menu.tsx
@@ -94,7 +94,7 @@ const Menu = () => {
                     <button
                         className={"bg-blue-700 hover:bg-blue-700/50 text-white font-bold py-2 px-4 ml-2 rounded"}
                         hidden={use_node_editor_only}
-                        onClick={toggle_info_panel}
+                        onClick={useObjectInfoStore.getState().toggle}
                     >{info_svg}</button>
                     <div className="relative">
                         {hasAnimation && <AnimationControls/>}
