+++ src/frontend/src/components/viewer/DynamicGridHelper.tsx
@@ -3,7 +3,7 @@ import React, { useMemo } from 'react';
 import { GridHelper as ThreeGridHelper } from 'three';
 import { useThree } from '@react-three/fiber';
 import { useModelStore } from '../../state/modelStore';
-import { calculateGridSize } from '../../utils/calculateGridSize';
+import { calculateGridSize } from '../../utils/scene/calculateGridSize';
 
 const DynamicGridHelper: React.FC = () => {
   const { boundingBox } = useModelStore();
