+++ src/frontend/src/components/viewer/ThreeModel.tsx
@@ -6,11 +6,12 @@ import * as THREE from 'three';
 import {GLTFResult, ModelProps} from '../../state/modelInterfaces';
 import {useAnimationStore} from '../../state/animationStore';
 import {useAnimationEffects} from '../../hooks/useAnimationEffects';
-import {handleMeshSelected} from '../../utils/mesh_handling';
 import {useModelStore} from '../../state/modelStore';
-import {replaceBlackMaterials} from '../../utils/assignDefaultMaterial';
-import {useTreeViewStore, generateTree} from '../../state/treeViewStore';
+import {replaceBlackMaterials} from '../../utils/scene/assignDefaultMaterial';
+import {useTreeViewStore} from '../../state/treeViewStore';
 import {useOptionsStore} from "../../state/optionsStore";
+import { generateTree } from '../../utils/tree_view/generateTree';
+import {handleClickMesh} from "../../utils/mesh_select/handleClickMesh";
 
 
 const ThreeModel: React.FC<ModelProps> = ({url}) => {
@@ -109,7 +110,7 @@ const ThreeModel: React.FC<ModelProps> = ({url}) => {
     return (
         <primitive
             object={scene}
-            onClick={handleMeshSelected}
+            onClick={handleClickMesh}
             dispose={null}
         />
     );
