+++ src/frontend/src/state/treeViewStore.ts
@@ -1,5 +1,4 @@
 import {create} from 'zustand';
-import * as THREE from 'three';
 
 export interface TreeNode {
     id: string;
@@ -26,20 +25,3 @@ export const useTreeViewStore = create<TreeViewState>((set) => ({
     isCollapsed: true,
     setIsCollapsed: (collapsed) => set({isCollapsed: collapsed}),
 }));
-
-// Utility function to generate tree data from the scene, excluding LineSegments
-export const generateTree = (object: THREE.Object3D): TreeNode | null => {
-    // Check if the object is a LineSegments and skip it
-    if (object instanceof THREE.LineSegments) {
-        return null;
-    }
-
-    return {
-        id: object.uuid,
-        name: object.name || object.type,
-        // Filter out null children (those that are LineSegments)
-        children: object.children
-            .map((child) => generateTree(child))
-            .filter((child): child is TreeNode => child !== null),
-    };
-};
