+++ src/frontend/src/utils/default_materials.ts
@@ -1,0 +1,4 @@
+import * as THREE from "three";
+
+export const selectedMaterial = new THREE.MeshStandardMaterial({color: 'blue', side: THREE.DoubleSide});
+export const defaultMaterial = new THREE.MeshStandardMaterial({color: 'white', side: THREE.DoubleSide});
