+++ src/frontend/src/utils/fb_handling/handle_incoming_buffers.ts
@@ -1,8 +1,8 @@
 import {CommandType, Message} from '../../flatbuffers/wsock'
 import * as flatbuffers from "flatbuffers";
 import {reply_ping} from "./reply_ping";
-import {update_scene_from_message} from "../scene/update_scene_from_message";
-import {receive_mesh_info_reply} from "../mesh_select/receive_mesh_info_reply";
+import {update_scene_from_message} from "../scene/comms/update_scene_from_message";
+import {receive_mesh_info_reply} from "../mesh_select/comms/receive_mesh_info_reply";
 import {update_nodes} from "../node_editor/update_nodes";
 import {handle_finished_procedure} from "../node_editor/handle_finished_procedure";
