--- src/frontend/src/utils/info_panel_utils.ts
@@ -1,5 +1,0 @@
-import {useObjectInfoStore} from "../state/objectInfoStore";
-
-export function toggle_info_panel() {
-    useObjectInfoStore.getState().toggle();
-}
