+++ src/frontend/src/utils/mesh_select/handleClickEmptySpace.ts
@@ -1,0 +1,6 @@
+import {deselectObject} from "./deselectObject";
+
+export function handleClickEmptySpace(event: MouseEvent) {
+    event.stopPropagation();
+    deselectObject();
+}
