+++ src/frontend/src/utils/scene/assignDefaultMaterial.ts
