+++ src/frontend/src/utils/scene/calculateGridSize.ts
