+++ src/frontend/src/utils/scene/comms/update_scene_from_message.ts
@@ -1,5 +1,5 @@
-import {Message} from "../../flatbuffers/wsock/message";
-import {useModelStore} from "../../state/modelStore";
+import {Message} from "../../../flatbuffers/wsock/message";
+import {useModelStore} from "../../../state/modelStore";
 
 export const update_scene_from_message = (message: Message) => {
     console.log('Received scene update message from server');
